feature/REG-60
- extract chart logic in separate js file.
